move our data structures into a central location
